requirements.txt
torch
torchvision
torchaudio
imutils
pytorch-fid
scikit-learn
xgboost
pandas
tqdm
tensorboardX
tensorboard
opencv-python
pillow
scipy
numpy
matplotlib